requirements.txt
fastapi
uvicorn
requests
google-cloud-bigquery
google-cloud-logging
pydantic
